Simulate Boreal Caribou population from survival, ageing and birth.

Simulate population projection for Boreal Caribou from key survival and fecundity rates.

bbs_population_caribou(
  survival,
  fecundity,
  adult_females = 1000,
  proportion_adult_female = 0.65,
  proportion_yearling_female = 0.5,
  stochastic = TRUE
)

Arguments

survival

A list of the Caribou survival rates (output of bbs_survival_caribou())

fecundity

A list of the Caribou fecundity rates (output of bbs_fecundity_caribou())

adult_females

A number of the initial number of adult females in the population.

proportion_adult_female

A number between 0 and 1 of the proportion of adults that are female.

proportion_yearling_female

A number between 0 and 1 of the proportion of yearlings that are female.

stochastic

A flag indicating whether to include demographic stochasticity.

Value

A matrix of the population by stage and period.

Details

This model assumes that survival occurs at the end of each period and survival, ageing and birth occur at the end of each year, in that order. Initial population size for all stages is determined from the initial number of adult females and the calculated stable stage distribution. Survival and fecundity arguments accept outputs of bbs_survival_caribou() and bbs_fecundity_caribou(). These are converted into processes matrices using bbs_matrix_survival_period() and bbs_matrix_birth_year() prior to projection. bbs_population() is called internally to project population. Yearling female survival is assumed to be the same as adult female survival. Stages 1-6 in output matrix correspond to female calf, male calf, female yearling, male yearling, female adult and male adult, respectively.

Examples

if (interactive()) {
    survival <- bbs_survival_caribou(0.84)
    fecundity <- bbs_fecundity_caribou(0.7)
    x <- bbs_population_caribou(survival, fecundity = fecundity)
}