Parameter Descriptions for bboutools Functions

Parameter Descriptions for bboutools Functions

Arguments

...

Unused parameters.

x

The object.

population_name

A string of the population name. This does not affect simulation but can be used as a unique identifier.

survival_adult_female

A number between 0 and 1 of the annual female adult survival.

survival_calf_female

A number between 0 and 1 of the annual female calf survival.

calves_per_adult_female

A number of the calves per adult female.

proportion_female

A number between 0 and 1 indicating the proportion of recruits that are female.

nyear

A whole number of the number of years.

nperiod_within_year

A whole number of the number of periods in a year.

female_recruit_stage

A positive whole number of the stage representing female recruits.

male_recruit_stage

A positive whole number of the stage representing male recruits. Ignored if NULL.

stochastic

A flag indicating whether to include demographic stochasticity.

month_collar

A whole number between 1 and 12 of the collaring month.

collared_adult_females

A whole positive number of the number of collared adult females. The number of collared adult females is 'topped up' each year at month_collar.

probability_uncertain_mortality

A number between 0 and 1 of the probability of uncertain mortality.

probability_uncertain_survival

A number between 0 and 1 of the probability of uncertain survival.

probability_unsexed_adult_female

A number between 0 and 1 of the probability that an adult female is unsexed.

probability_unsexed_adult_male

A number between 0 and 1 of the probability that an adult male is unsexed.

adult_females

A number of the initial number of adult females in the population.

proportion_adult_female

A number between 0 and 1 of the proportion of adults that are female.

proportion_yearling_female

A number between 0 and 1 of the proportion of yearlings that are female.

population

A matrix of the population by stage and period (output of bbs_population() or bbs_population_caribou()).

group_size_lambda

A number of the lambda value of the gamma-poisson distribution to draw group sizes from.

group_size_theta

A number of the theta value of the gamma-poisson distribution to draw group sizes from.

group_max_proportion

A number between 0 and 1 of the maximum group size as proportion of the total population.

group_min_size

A whole positive number of the minimum group size.

month_composition

A whole number between 1 and 12 of the month that composition surveys take place, relative to the start of the biological year.

group_coverage

A number between 0 and 1 of the proportion of groups sampled.

recruit_stages

A vector of whole numbers indicating the recruit stages (e.g., calf).

reproductive_female_stages

A vector of whole numbers indicating the reproductive female stages (e.g., cows).

survival

A list of the Caribou survival rates (output of bbs_survival_caribou())

fecundity

A list of the Caribou fecundity rates (output of bbs_fecundity_caribou())